backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
pytesseract
pillow
httpx
pydantic
python-dotenv
langchain
langchain-openai
langgraph
openai
google-genai
